% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bickeldocksum.default.R
\name{bickeldoksum.default}
\alias{bickeldoksum.default}
\title{Bickel-Doksum Default}
\usage{
\method{bickeldoksum}{default}(object, ...)
}
\arguments{
\item{object}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{\dots}{additional arguments to be passed to the estimation function;  see bickeldoksumEst()}
}
\value{
an object of class \code{transformation}; see bickeldoksumEst()
}
\description{
Bickel-Doksum estimation
}
\keyword{internal}

