% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bickeldocksum.formula.R
\name{bickeldoksum.formula}
\alias{bickeldoksum.formula}
\title{Bickel-Doksum Formula}
\usage{
\method{bickeldoksum}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{object of type formula with the model to transform}

\item{data}{an optional data frame containing the variables in the model}

\item{\dots}{additional arguments to be passed to the estimation function;  see bickeldoksumEst()}
}
\value{
an object of class \code{transformation}; see bickeldoksumEst()
}
\description{
Bickel-Doksum estimation
}
\keyword{internal}
