% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bickeldocksum.lm.R
\name{bickeldoksum.lm}
\alias{bickeldoksum.lm}
\title{Bickel-Docksum lm}
\usage{
\method{bickeldoksum}{lm}(object, ...)
}
\arguments{
\item{object}{of type lm with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see bickeldocksumEst()}
}
\value{
an object of class \code{transformation}; see bickeldocksumEst()
}
\description{
Bickel-Docksum estimation
}
\keyword{internal}
