% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bickeldocksum.lme.R
\name{bickeldoksum.lme}
\alias{bickeldoksum.lme}
\title{Bickel-Docksum lme}
\usage{
\method{bickeldoksum}{lme}(object, method, lambdarange = c(1e-11, 2),
  tol = 1e-04, ...)
}
\arguments{
\item{object}{of type lme with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see bickeldocksumEst()}
}
\value{
an object of class \code{transformation}; see bickeldocksumEst()
}
\description{
Bickel-Docksum estimation
}
\keyword{internal}
