% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bickeldocksum.R
\name{bickeldoksum}
\alias{bickeldoksum}
\title{Bickel-Doksum Transformation}
\usage{
bickeldoksum(object, lambda = seq(0.05, 2, 0.05), ...)
}
\arguments{
\item{object}{formula or fitted model; see below.}

\item{data}{an optional data frame, list or environment as in lm.}

\item{lambdarange}{range for the estimation parameter lambda - default c(0, 2)}
}
\value{
An object of class \code{transformation} with the following arguments.

llike The value of the profile log-likelihood at its maximum.

lambdahat The value of lambda for which the log-likelihood is maximized

logvector The profile log-likelihood evaluated at \code{lambdavector}.

lambdavector A sequence with optional values for \code{lambda}.

family Employed family of transformations.


zt Vector of the (normalized) transformed response variable \code{y}

yt Vector of the transformed response variable \code{y}.

modelt An object of type \code{lm} employing the transformed vector \code{zt} as the response variable.
}
\description{
Estimates the transformation parameter for the Bickel-Doksum transformation employing the profile log-likehood. Only accepts positive values for the dependent variable.
}
\references{
Bickel, P., and Doksum, K. 1981 An analysis of transformations revisited. \emph{Journal of the American Statistical Association}, \bold{76(374)}, 296--311.
}
\keyword{Bickel-Doksum}
