% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/blast_methods.R
\name{blast}
\alias{blast}
\alias{blast.data.frame}
\title{Split records of a data frame by variable values.}
\usage{
blast(x, ...)

\method{blast}{data.frame}(x, ..., .drop = TRUE)
}
\arguments{
\item{x}{a data frame or tibble.}

\item{...}{one or more unquoted variables used for splitting.}

\item{.drop}{logical, should empty levels of the vector be skipped from
the output? Defaults to TRUE.}
}
\description{
Splits records of a data frame by values of its variable
or variables. If the splitting factor is a single variable, this operation
reverts the effects of \code{\link{compress}}.
}
\details{
S3 generic function.
}
