% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{blast}
\alias{blast}
\title{Split records of a data frame by variable values.}
\usage{
blast(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Splits records of a data frame by values of its variable
or variables. If the splitting factor is a single variable, this operation
reverts the effects of \code{\link{compress}}.
}
