% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafos.R
\name{box_cox}
\alias{box_cox}
\title{Transformation: Box Cox}
\usage{
box_cox(y, lambda = lambda, shift = 0)
}
\arguments{
\item{y}{vector of response variables}

\item{lambda}{transformation parameter}

\item{shift}{bla bla}
}
\value{
Depending on the selected \code{method} the return is a log
likelihood, a skewness, a pooled skewness or a Kolmogorov-Smirnoff, Craemer
von Mises or Kullback Leibner divergence.
}
\description{
Transformation: Box Cox
}
\keyword{internal}
