% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_cx.default.R
\name{bx_cx.default}
\alias{bx_cx.default}
\title{Box-Cox Default}
\usage{
\method{bx_cx}{default}(object, ...)
}
\arguments{
\item{object}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{\dots}{additional arguments to be passed to the estimation function;  see bcxEst()}
}
\value{
an object of class \code{transformation}; see bcxEst()
}
\description{
Box-Cox estimation
}
\keyword{internal}
