% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_cx.formula.R
\name{bx_cx.formula}
\alias{bx_cx.formula}
\title{Box-Cox Formula}
\usage{
\method{bx_cx}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{object of type formula with the model to transform}

\item{data}{an optional data frame containing the variables in the model}

\item{\dots}{additional arguments to be passed to the estimation function; see bcxEst()}
}
\value{
an object of class \code{transformation}; see bcxEst()
}
\description{
Box-Cox estimation
}
\keyword{internal}

