% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_cx.lm.R
\name{bx_cx.lm}
\alias{bx_cx.lm}
\title{Box-Cox lm}
\usage{
\method{bx_cx}{lm}(object, method, lambdarange = c(-2, 2), tol = 1e-04, ...)
}
\arguments{
\item{object}{of type lm with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see bcxEst()}
}
\value{
an object of class \code{transformation}; see bcxEst()
}
\description{
Box-Cox estimation
}
\keyword{internal}
