% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_cx.lme.R
\name{bx_cx.lme}
\alias{bx_cx.lme}
\title{Box-Cox lme}
\usage{
\method{bx_cx}{lme}(object, ...)
}
\arguments{
\item{object}{of type lme with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see bcxEst()}
}
\value{
an object of class \code{transformation}; see bcxEst()
}
\description{
Box-Cox estimation
}
\keyword{internal}
