% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_cx.R
\name{bx_cx}
\alias{bx_cx}
\title{Box-Cox Transformation}
\usage{
bx_cx(object, ...)
}
\arguments{
\item{object}{formula or fitted model; see below.}

\item{data}{an optional data frame, list or environment as in lm.}

\item{lambdarange}{range for the estimation parameter lambda - c(-2,2)}
}
\value{
An object of class \code{transformation} with the following arguments

llike The value of profile log-likelihood at its maximum

lambdahat The value of lambda for which the log-likelihood is maximized

logvector The profile log-likelihood evaluated at \code{lambdavector}

lambdavector A sequence with optional values for \code{lambda}

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

zt Vector of the (normalized) transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{zt} as the response variable
}
\description{
Estimates the transformation parameter for the Box-Cox transformation employing the profile log-likehood. Only accepts positive values for the dependent variable.
}
\references{
Box, G. E. P., and Cox, D. R. 1964 An analysis of transformations. \emph{Journal of the Royal Statistical Society}, \bold{26(2)}, 211--252.
}
\keyword{boxcox}
