% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_cxEst.R
\name{bx_cxEst_lm}
\alias{bx_cxEst_lm}
\title{Box Cox Estimation - lm}
\usage{
bx_cxEst_lm(y, x, method, lambdarange, tr = FALSE, ...)
}
\arguments{
\item{y}{vector of response variables}

\item{x}{matrix of regressors}

\item{lambdarange}{range for the estimation parameter expr(lambda) - default c(-2, 2)}

\item{tr}{logical value. if tr = TRUE warning messages for the likelihood functions are suppressed - default FALSE}
}
\value{
An object of class \code{transformation} with the following arguments

llike The value of \code{profile log-likelihood} at its maximum

logvector The profile log-likelihood evaluated at \code{lambdavector}

lambdavector Employed family of transformations

A sequence with optional values for \code{lambda}

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Box Cox Estimation - lm
}
\keyword{internal}
