% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTransformation.default.R
\name{compareTransformation.default}
\alias{compareTransformation.default}
\title{Compare Transformation Default}
\usage{
\method{compareTransformation}{default}(object, ...)
}
\arguments{
\item{object}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{\dots}{additional arguments to be passed to the estimation function; see docompareTransformation()}
}
\value{
an object of class \code{transformation}; see docompareTransformation()
}
\description{
Compares families of transformations with respect to  sense of AIC, R.Squared, Skewness or Kurtosis; For positive y compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum
}
\keyword{internal}

