% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTransformation.lm.R
\name{compareTransformation.lm}
\alias{compareTransformation.lm}
\title{Compare Transformation lm}
\usage{
\method{compareTransformation}{lm}(object, ...)
}
\arguments{
\item{object}{a object of type lm with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see docompareTransformation()}
}
\value{
an object of class \code{transformation}; see docompareTransformation()
}
\description{
Compares families of transformations with respect to  sense of AIC, R.Squared, Skewness or Kurtosis; For positive y compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum
}
\keyword{internal}
