% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/compress_methods.R
\name{compress}
\alias{compress}
\alias{compress.default}
\alias{compress.list}
\title{Convert a named object into a tibble.}
\usage{
compress(x, names_to = "names", ...)

\method{compress}{default}(x, names_to = "names", values_to = "values", ...)

\method{compress}{list}(x, names_to = "names", values_to = "values", simplify = TRUE, ...)
}
\arguments{
\item{x}{a named object, e.g. atomic vector, to be converted to a tibble.}

\item{names_to}{name of the variable storing the names.}

\item{...}{extra arguments, currently none.}

\item{values_to}{name of the variable storing the object's elements.}

\item{simplify}{logical, should a simplified output be considered for a
list of compatible data frames? See the details section. Default is TRUE.}
}
\value{
a tibble.
}
\description{
Converts a named object, e.g. an atomic vector or a list
into a tibble.
The object elements are stored in a variable named as specified by the
'values_to' argument, the names are stored under 'names_to' variable.
}
\details{
By setting 'simplify' to TRUE, a special behavior for a 'pure' list
of data frames is toggled on. The function detects if all data frames have
identical column names. If so, the result is coerced into a single tibble or
data frame and the values_to argument is ignored.
Otherwise or if simplify is set to FALSE, each element is of the list is
stored in the variable specified by the values_to argument.
The argument simplify set to FALSE makes the function faster.
`compress()` is a S3 generic function. Its effect can by inverted e.g.
by applying \code{\link[plyr]{dlply}} or \code{\link{blast}}.
}
