% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_methods.R
\name{compress.list}
\alias{compress.list}
\title{Convert a named list into a tibble.}
\usage{
\method{compress}{list}(x, names_to = "names", values_to = "values", simplify = TRUE, ...)
}
\arguments{
\item{x}{a named list, to be converted to a tibble.}

\item{names_to}{name of the variable storing the names.}

\item{values_to}{name of the variable storing the object's elements.}

\item{simplify}{logical, should a simplified output be considered for a
list of compatible data frames? See the details section. Default is TRUE.}

\item{...}{extra arguments, currently none.}
}
\value{
a tibble.
}
\description{
Converts a named list into a tibble.
The object elements are stored in a variable named as specified by the
'values_to' argument, the names are stored under 'names_to' variable.
}
\details{
By setting 'simplify' to TRUE, a special behavior for a 'pure' list
of data frames is toggled on. The function detects if all data frames have
identical column names. If so, the result is coerced into a single tibble or
data frame and the values_to argument is ignored.
Otherwise or if simplify is set to FALSE, each element is of the list is
stored in the variable specified by the values_to argument.
The argument simplify set to FALSE makes the function faster.
}
