% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{compress}
\alias{compress}
\title{Compact a vector or a list into a data frame.}
\usage{
compress(x, names_to = "names", ...)
}
\arguments{
\item{x}{an object.}

\item{names_to}{name of the tibble column storing the names of
the input object x.}

\item{...}{extra arguments passed to the methods.}
}
\value{
a tibble.
}
\description{
Coverts a named vector or a list to a tibble. Names of the
vector are stored in a variable specified by 'names_to'.
}
\details{
a S3 generic function. Its effect can by inverted e.g. by applying
\code{\link[plyr]{dlply}}.
}
