% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{data_transformation}
\alias{data_transformation}
\title{Tranforms dependent variables}
\usage{
data_transformation(fixed, smp_data, transformation, lambda)
}
\arguments{
\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the nested error linear regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators. The argument corresponds
to the argument \code{fixed} in function \code{\link[nlme]{lme}}.}

\item{smp_data}{a data frame that needs to comprise all variables named in
\code{fixed}. If transformed data is further used to fit a nested error
linear regression model \code{smp_data} also needs to comprise the variable
named in \code{smp_domains} (see \code{\link{ebp}}).}

\item{transformation}{a character string. Three different transformation
methods for the dependent variable can be chosen (i) no transformation ("no");
(ii) log transformation ("log"); (iii) Box-Cox transformation
("box.cox").}

\item{lambda}{a number that determines the Box-Cox transformation. In case of
no and log transformation \code{lambda} can be set to NULL.}
}
\value{
a named list with two elements, a data frame containing the
transformed data set (\code{transformed_data}) and a shift parameter
\code{shift} if present. In case of no transformation the original data
frame is returned and the shift parameter is NULL.
}
\description{
Function \code{data_transformation} is a wrapper function that provides
three different transformation types for the dependent variable in a
data set.
}
\examples{
# Loading data
data("Xoutsamp_AuxVar")
data("incomedata")

# Determine the optimal parameter
opt_par <- optimal_parameter(generic_opt, income~educ1, incomedata,
"provlab", "box.cox")

# Transform sample data
transform_data <- data_transformation(income~educ1, incomedata, "box.cox",
lambda=opt_par)

# Fit the nested error linear regression model
mixed_model <- lme(income~educ1, transform_data$transformed_data,
random=~1|provlab)
}
\seealso{
\code{\link{optimal_parameter}}, \code{\link[nlme]{lme}}
}

