% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.compare_trafo.R
\name{diagnostics.compare_trafo}
\alias{diagnostics.compare_trafo}
\title{Diagnostics for two differently transformed models}
\usage{
\method{diagnostics}{compare_trafo}(object, ...)
}
\arguments{
\item{object}{an object of type \code{compare_trafo}}

\item{...}{additional arguments that are not used in this method}
}
\value{
an object of class \code{summary.transformation}
}
\description{
Information about the transformed data and model and components of an 
transformation object are extracted. The returned object is suitable for 
printing with the print.summary.transformation method.
}
