% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostics for fitted models}
\usage{
diagnostics(object, ...)
}
\arguments{
\item{object}{an object that contains two models that should be compared.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
The return depends on the class of its argument. The 
documentation of particular methods gives detailed information about the 
return of that method.
}
\description{
Returns information about the transformation and selected diagnostics to 
check model assumptions.
}
\seealso{
\code{\link{diagnostics.trafo_lm}}, 
\code{\link{diagnostics.trafo_compare}}
}
\keyword{diagnostics}
