% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_internal.R
\name{diagnostics_internal}
\alias{diagnostics_internal}
\title{Internal diagnostic functions}
\usage{
diagnostics_internal(modOne, modTwo)
}
\description{
The function is called in diagnostics.trafo_lm and 
diagnostics.trafo_compare.
}
\keyword{internal}
