% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estims.R
\name{divergence_min_CvM}
\alias{divergence_min_CvM}
\title{Divergence minimization by Cramer von Mises}
\usage{
divergence_min_CvM(res = res)
}
\arguments{
\item{res}{residuals from a linear model with response variable yt and 
explanatory variables x}
}
\value{
sum of squared differences
}
\description{
Divergence minimization by Cramer von Mises
}
\keyword{internal}
