% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estims.R
\name{divergence_min_KL}
\alias{divergence_min_KL}
\title{Divergence minimization by Kullback-Leibler}
\usage{
divergence_min_KL(res = res)
}
\arguments{
\item{res}{residuals from a linear model with response variable yt and 
explanatory variables x}
}
\value{
median of Kullback-Leibler divergence
}
\description{
Divergence minimization by Kullback-Leibler
}
\keyword{internal}
