% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estims.R
\name{divergence_min_KS}
\alias{divergence_min_KS}
\title{Divergence minimization by Kolmogorov Smirnov}
\usage{
divergence_min_KS(res = res)
}
\arguments{
\item{res}{residuals from a linear model with response variable yt and 
explanatory variables x}
}
\value{
differences of supremum
}
\description{
Divergence minimization by Kolmogorov Smirnov
}
\keyword{internal}
