% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoCompareTransformation.R
\name{docompareTransformation}
\alias{docompareTransformation}
\title{Selects Transformation}
\usage{
docompareTransformation(data, lambda.dual = 0.5, plotit = TRUE, ...)
}
\arguments{
\item{data}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}
}
\value{
An object of class \code{transformation} with the following arguments

llike The value of profile log-likelihood at its maximum

logvector The profile log-likelihood evaluated at \code{lambdavector}

lambdavector Employed family of transformations

A sequence with optional values for \code{lambda}

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Chooses the 'best' transformation in the sense of AIC, R.Squared, Skewness or Kurtosis; For positive \code{y} compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum
}
\keyword{internal}

