% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doselectTransformation.R
\name{doselectTransformation}
\alias{doselectTransformation}
\title{Selects Transformation}
\usage{
doselectTransformation(data, criterion = "AIC", usefamily = NULL, ...)
}
\arguments{
\item{criterion}{a character string to select the desired criterion to be used to select a family of transformation whose parameter is estimated using the likelihood approach. The available options ara (i) Akaike's information criterion "AIC"; (ii) R.Squared "R2"; (iii) Skewnnes "SK"; (iv) Kurtosis "KT".}

\item{usefamily}{a character string. Six different transformation families for the dependent variable can be chosen. For positive response variable the available options are (i) "Box-Cox" or "BC" ; (ii) "Dual" or "D" ("log"); (iii) "Manly" or "Ma". For negative response variables (i) "Yeo-Johhson" or "YJ"; (ii) "Modulus" or "Mod"; (iii) "Bickel-Doksum" or "BD"; (iv) "Manly" or "Ma"}

\item{object}{an optional data frame, list or environment as in lm.}
}
\value{
An object of class \code{transformation} with the following arguments

llike The value of profile log-likelihood at its maximum

logvector The profile log-likelihood evaluated at \code{lambdavector}

lambdavector A sequence with optional values for \code{lambda}

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Chooses the 'best' transformation in the sense of AIC, R.Squared, Skewness or Kurtosis; For positive \code{y} compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum
}
\keyword{internal}
