% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.default.R
\name{dual.default}
\alias{dual.default}
\title{Dual Default}
\usage{
\method{dual}{default}(object, ...)
}
\arguments{
\item{object}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{\dots}{additional arguments to be passed to the estimation function;  see dualEst()}
}
\value{
an object of class \code{transformation}; see dualEst()
}
\description{
Dual estimation
}
\keyword{internal}

