% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.formula.R
\name{dual.formula}
\alias{dual.formula}
\title{Dual Formula}
\usage{
\method{dual}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{object of type formula with the model to transform}

\item{data}{an optional data frame containing the variables in the model}

\item{\dots}{additional arguments to be passed to the estimation function; see dualEst()}
}
\value{
an object of class \code{transformation}; see dualEst()
}
\description{
Dual estimation
}
\keyword{internal}

