% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.lm.R
\name{dual.lm}
\alias{dual.lm}
\title{Dual lm}
\usage{
\method{dual}{lm}(object, ...)
}
\arguments{
\item{object}{of type lm with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see dualEst()}
}
\value{
an object of class \code{transformation}; see dualEst()
}
\description{
Dual estimation
}
\keyword{internal}
