% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.lme.R
\name{dual.lme}
\alias{dual.lme}
\title{Dual lme}
\usage{
\method{dual}{lme}(object, method, lambdarange = c(0, 2), tol = 1e-04, ...)
}
\arguments{
\item{object}{of type lme with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see dualEst()}
}
\value{
an object of class \code{transformation}; see dualEst()
}
\description{
Dual estimation
}
\keyword{internal}
