% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{dual}
\alias{dual}
\title{Dual Transformation}
\usage{
dual(object, ...)
}
\arguments{
\item{object}{formula or fitted model; see below.}

\item{data}{an optional data frame, list or environment as in lm.}


\item{lambda.dual}{a value to be used as the parameter of the dual transformation - default 0.5.}
}
\value{
An object of class \code{transformation} with the following arguments

lambda.user Value for \code{lambda} introduced by the user

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

zt Vector of the (normalized) transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{zt} as the response variable
}

\references{
Yang, Z. 2004 A modified family of power transformations. \emph{Research Collection School Of Economics}
}
\description{
Transformed a response variable for a given \code{lambda} using the Dual transformation. Only accepts positive values for the dependent variable.
}
