% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafoEst.R
\name{est_lm}
\alias{est_lm}
\title{Estimation of optimal transformation parameter - lm}
\usage{
est_lm(y, x, method, lambdarange, trafo, custom_func, custom_func_std, ...)
}
\arguments{
\item{y}{vector of response variables}

\item{x}{matrix of regressors}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Maximum likelihood approach ("ml"), (ii) Skewness minimization ("skew"),
(iii) Kurtosis optimization ("kurt"), (iv) Divergence minimization by 
Kolmogorov-Smirnov ("div.ks"), by Cramer-von-Mises ("div.cvm") or by 
Kullback-Leibler ("div.kl"). Defaults to "ml".}

\item{lambdarange}{range for the estimation parameter expr(lambda) - default c(-2, 2)}
}
\value{
modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Estimation of optimal transformation parameter - lm
}
\keyword{internal}
