% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafoEst.R
\name{est_lme}
\alias{est_lme}
\title{Box Cox Estimation - lme}
\usage{
est_lme(y, x, formula, data, rand_eff, method = method,
  lambdarange = lambdarange, trafo, custom_func, custom_func_std)
}
\arguments{
\item{y}{vector of response variables}

\item{x}{matrix of regressors}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Restricted maximum likelihood approach ("reml"), 
(ii) Skewness minimization ("skew") and pooled skewness minimization ("pskew"), 
(iii) Divergence minimization by Kolmogorov-Smirnov ("div.ks"), 
by Cramer-von-Mises ("div.cvm") or by Kullback-Leibler ("div.kl").}

\item{lambdarange}{range for the estimation parameter expr(lambda) - default c(-2, 2)}
}
\value{
modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Box Cox Estimation - lme
}
\keyword{internal}
