% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eusilcA_Vienna.R
\docType{data}
\name{eusilcA_Vienna}
\alias{eusilcA_Vienna}
\title{Simulated eusilc data - for state Vienna}
\format{A data frame with 236 observations and 14 variables.}
\usage{
eusilcA_Vienna
}
\description{
The data set is synthetic EU-SILC data based on \code{\link[emdi]{eusilcA_smp}} of package \bold{emdi}.
It contains the observations from state Vienna.
The data set contains 14 variables for the state Vienna. The variables are as follows:
\itemize{
  \item eqIncome. (numeric) a simplified version of the equivalized household income.
  \item eqsize. (numeric) the equivalized household size according to the modified OECD scale.
  \item cash. (numeric) employee cash or near cash income (net).
   \item  self_empl. (numeric) cash benefits or losses from self-employment (net).
  \item  unempl_ben. (numeric) unemployment benefits (net).
   \item age_ben.  (numeric) old-age benefits (net).
   \item surv_ben (numeric) survivor's benefits (net).
   \item  sick_ben (numeric) sickness benefits (net).
  \item dis_ben (numeric) disability benefits (net).
   \item rent (numeric) income from rental of a property or land (net).
  \item fam_allow (numeric)  family/children related allowances (net).
  \item  house_allow (numeric)  housing allowances (net).
 \item cap_inv (numeric)  interest, dividends, profit from capital investments in unincorporated business (net).
\item tax_adj (numeric)  repayments/receipts for tax adjustment (net).bles.
 ...
}
}
\keyword{datasets}
