% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_methods.R
\name{exchange.data.frame}
\alias{exchange.data.frame}
\title{Substitute/exchange elements of a data frame variable.}
\usage{
\method{exchange}{data.frame}(x, variable, dict, key = "variable", value = "label", ...)
}
\arguments{
\item{x}{a data frame, storing the variable to be substituted.}

\item{variable}{name of the variable to be substituted.}

\item{dict}{a data frame with the recoding scheme. It has to contain columns
specified by the key and value arguments.}

\item{key}{the name of the dict data frame variable corresponding to
the elements of x: the search key.}

\item{value}{the name of the dict data frame variable containing the elements
used for substitution of x: the returned labels/values.}

\item{...}{extra arguments, currently none.}
}
\value{
a vector.
}
\description{
Substitutes elements of a data frame variable with values
provided by a dictionary: a data frame with the re-coding scheme.
}
\details{
Any numeric variable is silently coerced
to a character one.
}
