% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/exchange_methods.R
\name{exchange}
\alias{exchange}
\alias{exchange.default}
\alias{exchange.list}
\alias{exchange.data.frame}
\title{Substitute/exchange object elements.}
\usage{
exchange(x, ...)

\method{exchange}{default}(x, dict, key = "variable", value = "label", ...)

\method{exchange}{list}(x, dict, key = "variable", value = "label", ...)

\method{exchange}{data.frame}(x, variable, dict, key = "variable", value = "label", ...)
}
\arguments{
\item{x}{an object, e.g an atomic vector, matrix, list or a data frame.}

\item{...}{extra arguments, currently none.}

\item{dict}{a data frame with the re-coding scheme. It has to contain columns
specified by the key and value arguments.}

\item{key}{the name of the dict data frame variable corresponding to
the elements of x: the search key.}

\item{value}{the name of the dict data frame variable containing the elements
used for substitution of x: the returned labels/values.}

\item{variable}{a quoted or unquoted variable
of the data frame to be re-coded.}
}
\value{
a vector, matrix, list or data frame, depending on `x`
}
\description{
Substitutes elements of a vector or other object with values
provided by a dictionary: a data frame with the re-coding scheme.
}
\details{
Any numeric object is silently coerced
to a character vector.
`exchange()` is a S3 generic function.
}
