% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_methods.R
\name{exchange.list}
\alias{exchange.list}
\title{Substitute/exchange elements of a list.}
\usage{
\method{exchange}{list}(x, dict, key = "variable", value = "label", ...)
}
\arguments{
\item{x}{a list, storing the elements to be substituted.}

\item{dict}{a data frame with the recoding scheme. It has to contain columns
specified by the key and value arguments.}

\item{key}{the name of the dict data frame variable corresponding to
the elements of x: the search key.}

\item{value}{the name of the dict data frame variable containing the elements
used for substitution of x: the returned labels/values.}

\item{...}{extra arguments. Those can be parameters passed to specific
exchange methods, e.g. variable argument passed to
\code{\link{exchange.data.frame}}, if a list of data frames is handled.}
}
\value{
a list.
}
\description{
Substitutes elements of a list with values
provided by a dictionary: a data frame with the re-coding scheme.
}
\details{
Numeric elements are silently coerced
to character.
}
