% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{exchange}
\alias{exchange}
\title{Exchange/substitute vector of data frame variable elements.}
\usage{
exchange(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to the methods.}
}
\value{
a vector or a data frame.
}
\description{
Substitutes elements of a vector or a data frame column
with values provided by a dictionary: a data frame with the re-coding scheme.
}
\details{
a S3 generic function.
}
