% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_methods.R
\name{full_rbind.data.frame}
\alias{full_rbind.data.frame}
\alias{full_rbind.matrix}
\alias{full_rbind.default}
\title{Full row binding of two data frames, matrices or vectors.}
\usage{
\method{full_rbind}{data.frame}(x, y, ...)

\method{full_rbind}{matrix}(x, y, ...)

\method{full_rbind}{default}(x, y, ...)
}
\arguments{
\item{x}{a data frame, tibble, matrix or vector.}

\item{y}{a data frame, tibble, matrix or vector.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame, a tibble or a matrix.
}
\description{
Binds two data frames, tibbles, matrices or vectors by rows,
missing variables are filled with NA.
}
\details{
Binding of objects of the same class allowed.
If a data frame binds with a tibble, the result is a data frame.
The binding partners need to be named (i.e. have column or vector names).
}
