% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_parameter.R
\name{generic_opt}
\alias{generic_opt}
\title{Selects estimation method for optimal parameter in Box-Cox transformations}
\usage{
generic_opt(lambda, fixed, smp_data, smp_domains, transformation, method)
}
\arguments{
\item{lambda}{a number that determines the Box-Cox transformation.}

\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the nested error linear regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators. The argument corresponds
to the argument \code{fixed} in function \code{\link[nlme]{lme}}.}

\item{smp_data}{a data frame that needs to comprise all variables named in
\code{fixed} and \code{smp_domains}.}

\item{smp_domains}{a character string containing the name of a variable
that indicates domains in the sample data. The variable can be numeric or a
factor.}

\item{transformation}{a character string. Three different transformation
types for the dependent variable can be chosen (i) no transformation ("no");
(ii) log transformation ("log"); (iii) Box-Cox transformation ("box.cox").}

\item{method}{a character string. In order to determine the optimal parameter
for the transformation six different estimation methods can be chosen
(i) restricted Maximum-Likelihood ("reml"); (ii) pooled skewness
minimization ("p.skew"); (iii) skewness minimization ("skew");
(iv) minimization of Kolmogorov-Smirnoff divergence  ("div.ks");
(v) minimization of Craemer von Mises divergence ("div.cvm"); (vi)
minimization of Kullback Leibner divergence  ("div.kl"). In case of no and
log transformation "NA" can be selected since no optimization is neccessary
for these two transformation types.}
}
\value{
Depending on the selected \code{method} the return is a log
likelihood, a skewness, a pooled skewness or a Kolmogorov-Smirnoff, Craemer
von Mises or Kullback Leibner divergence.
}
\description{
Function \code{generic_opt} selects the estimation method for the optimal
lambda that is neccessary to conduct Box-Cox transformations.
}
\examples{
# Example for method reml
log_likelihood <- generic_opt(0.0589, income~educ1, incomedata, "provlab",
"box.cox", "reml")
}
