% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modelt.R
\name{get_modelt}
\alias{get_modelt}
\title{Function that fits model with transformed dependent variable}
\usage{
get_modelt(object, trans_mod, std)
}
\arguments{
\item{object}{a model of type lm or lme}

\item{trans_mod}{the return of one of the transformation functions}

\item{std}{logical. If TRUE, the standardised transformations are used.}
}
\value{
A lm or lme model that uses the transformed variable as dependent 
variable.
}
\description{
Function that fits model with transformed dependent variable
}
\keyword{internal}
