% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transformed.R
\name{get_transformed}
\alias{get_transformed}
\title{Function that bundles the return of a trafo object}
\usage{
get_transformed(trafo, ans, y, lambda, custom_func, custom_func_std,
  custom_family)
}
\description{
Function that bundles the return of a trafo object
}
\keyword{internal}
