% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/rbind_methods.R
\name{inner_rbind}
\alias{inner_rbind}
\alias{inner_rbind.data.frame}
\alias{inner_rbind.matrix}
\alias{inner_rbind.default}
\title{Inner row binding of two data frames.}
\usage{
inner_rbind(x, y, ...)

\method{inner_rbind}{data.frame}(x, y, ...)

\method{inner_rbind}{matrix}(x, y, ...)

\method{inner_rbind}{default}(x, y, ...)
}
\arguments{
\item{x}{a data frame, tibble, matrix or vector.}

\item{y}{a data frame, tibble, matrix or vector.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame, tibble or a matrix.
}
\description{
Binds two data frames, tibbles, matrices or vectors by rows,
variables missing in one of the binding partners are removed.
}
\details{
Binding of objects of the same class allowed.
If a data frame binds with a tibble, the result is a data frame.
The binding partners need to be named (i.e. have column or vector names).
`inner_rbind()` is a S3 generic function.
}
