% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{inner_rbind}
\alias{inner_rbind}
\title{Inner row binding.}
\usage{
inner_rbind(x, y, ...)
}
\arguments{
\item{x}{an object.}

\item{y}{an object.}

\item{...}{additional arguments passed to methods.}
}
\value{
a data frame, a tibble or a matrix.
}
\description{
Binds two data frames, matrices or vectors by rows,
missing variables are removed.
}
