% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estims.R
\name{kurtosis_min}
\alias{kurtosis_min}
\title{Kurtosis}
\usage{
kurtosis_min(res = res)
}
\arguments{
\item{res}{residuals from a linear model with response variable yt and 
explanatory variables x}
}
\value{
absolute value of the skewness of the residuals
}
\description{
Kurtosis
}
\keyword{internal}
