% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logshift.R
\name{logshift}
\alias{logshift}
\title{Log shift transformation for linear models}
\usage{
logshift(object)
}
\arguments{
\item{object}{an object of type lm.}
}
\value{
an object of class \code{trafo}.
}
\description{
The function transforms the dependent variable of a linear model using the 
Log shift transformation. The shift parameter is determined by y + shift > 0.
}
\keyword{internal}
