% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manly.default.R
\name{manly.default}
\alias{manly.default}
\title{Manly Default}
\usage{
\method{manly}{default}(object, ...)
}
\arguments{
\item{object}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{\dots}{additional arguments to be passed to the estimation function;  see manlyEst()}
}
\value{
an object of class \code{transformation}; see manlyEst()
}
\description{
Manly Estimation
}
\keyword{internal}

