% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manly.formula.R
\name{manly.formula}
\alias{manly.formula}
\title{Manly Formula}
\usage{
\method{manly}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{object of type formula with the model to transform}

\item{data}{an optional data frame containing the variables in the model}

\item{\dots}{additional arguments to be passed to the estimation function; see manlyEst()}
}
\value{
an object of class \code{transformation}; see manlyEst()
}
\description{
Manly estimation
}
\keyword{internal}
