% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manly.lme.R
\name{manly.lme}
\alias{manly.lme}
\title{Manly lme}
\usage{
\method{manly}{lme}(object, method, lambdarange = c(-2, 2), tol = 1e-04,
  ...)
}
\arguments{
\item{object}{of type lme with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see manlyEst()}
}
\value{
an object of class \code{transformation}; see manlyEst()
}
\description{
Manly estimation
}
\keyword{internal}
