% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manly.R
\name{manly}
\alias{manly}
\title{Manly Transformation}
\usage{
manly(object, ...)
}
\arguments{
\item{object}{formula or fitted model; see below.}

\item{data}{an optional data frame, list or environment as in lm.}

\item{lambdarange}{range for the estimation parameter lambda - default c(-2, 2)}
}
\value{
An object of class \code{transformation} with the following arguments.

llike The value of profile log-likelihood at its maximum.

lambdahat The value of lambda for which the log-likelihood is maximized

logvector The profile log-likelihood evaluated at \code{lambdavector}.

lambdavector A sequence with optional values for \code{lambda}.

family Employed family of transformations.

yt Vector of the transformed response variable \code{y}.

zt Vector of the (normalized) transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{zt} as the response variable.
}
\description{
Estimates the transformation parameter for the Manly transformation employing the profile log-likehood.
}
\references{
Manly, B. F. J. 1976 Exponential data transformations. \emph{Journal of the Royal Statistical Society}, \bold{25(1)}, 37--42.
}
\keyword{Manly}
