% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manlyEst.R
\name{manlyEst}
\alias{manlyEst}
\title{Manly Estimation}
\usage{
manlyEst(y, x, lambdarange = c(-2, 2), tr = FALSE, ...)
}
\arguments{
\item{y}{vector of response variables}

\item{x}{matrix of regressors}

\item{lambdarange}{range for the estimation parameter expr(lambda) - default c(-2,2)}

\item{tr}{logical value. if tr = TRUE warning messages for the likelihood functions are suppressed - default FALSE}
}
\value{
An object of class \code{transformation} with the following arguments

llike The value of \code{profile log-likelihood} at its maximum

logvector The profile log-likelihood evaluated at \code{lambdavector}

lambdavector A sequence with optional values for \code{lambda}

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Manly Estimation
}
\keyword{internal}
