% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estims.R
\name{ML}
\alias{ML}
\title{Maximum Likelihood}
\usage{
ML(y, x, lambda, trafo, custom_func_std)
}
\arguments{
\item{y}{vector of response variables}

\item{x}{matrix of regressors}

\item{lambda}{transformation parameter}

\item{trafo}{determines the chosen transformation}

\item{custom_func_std}{a function that determines a customized standard
transformation.}
}
\value{
log-likelihood
}
\description{
Maximum Likelihood
}
\keyword{internal}
