% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.default.R
\name{modulus.default}
\alias{modulus.default}
\title{Modulus Default}
\usage{
\method{modulus}{default}(object, ...)
}
\arguments{
\item{object}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{\dots}{additional arguments to be passed to the estimation function;  see modulusEst()}
}
\value{
an object of class \code{transformation}; see modulusEst()
}
\description{
Modulus estimation
}
\keyword{internal}

