% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.formula.R
\name{modulus.formula}
\alias{modulus.formula}
\title{Modulus Formula}
\usage{
\method{modulus}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{object of type formula with the model to transform}

\item{data}{an optional data frame containing the variables in the model}

\item{\dots}{additional arguments to be passed to the estimation function;  see  modulusEst()}
}
\value{
an object of class \code{transformation}; see modulusEst()
}
\description{
Modulus estimation
}
\keyword{internal}

