% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.lm.R
\name{modulus.lm}
\alias{modulus.lm}
\title{Modulus lm}
\usage{
\method{modulus}{lm}(object, ...)
}
\arguments{
\item{object}{of type lm with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see modulusEst()}
}
\value{
an object of class \code{transformation}; see modulusEst()
}
\description{
Modulus estimation
}
\keyword{internal}

