% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.lme.R
\name{modulus.lme}
\alias{modulus.lme}
\title{Modulus lme}
\usage{
\method{modulus}{lme}(object, method, lambdarange = c(-2, 2), ...)
}
\arguments{
\item{object}{of type lme with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see modulusEst()}
}
\value{
an object of class \code{transformation}; see modulusEst()
}
\description{
Modulus estimation
}
\keyword{internal}
