% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{modulus}
\alias{modulus}
\title{Modulus Transformation}
\usage{
modulus(object, ...)
}
\arguments{
\item{object}{an object of type lm or lme.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
The return depends on the class of its argument. The 
documentation of particular methods gives detailed information about the 
return of that method.
}
\description{
Transforms the dependent variable of a regression model using the Modulus
transformation. The regression model can be a linear or a linear mixed model 
with one random intercept.
}
\seealso{
\code{\link{modulus.lm}},  \code{\link{modulus.lme}}
}
\keyword{Modulus}
