% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneparam.R
\name{oneparam}
\alias{oneparam}
\title{One parameter transformations}
\usage{
oneparam(object, ...)
}
\arguments{
\item{object}{an object of type lm or lme.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
The return depends on the class of its argument. The 
documentation of particular methods gives detailed information about the 
return of that method.
}
\description{
Transforms the dependent variable of a regression model using one parameter
transformations.
}
\seealso{
\code{\link{oneparam.lm}}
}
\keyword{internal}
