% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_parameter.R
\name{optimal_parameter}
\alias{optimal_parameter}
\title{Estimates optimal parameter for Box-Cox transformations}
\usage{
optimal_parameter(generic_opt, fixed, smp_data, smp_domains, transformation,
  method, interval = c(-1, 2))
}
\arguments{
\item{generic_opt}{a wrapper function that conducts the selected optimization
\code{method} in order to specify the optimal lambda for Box-Cox
transformations.}

\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the nested error linear regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators. The argument corresponds
to the argument \code{fixed} in function \code{\link[nlme]{lme}}.}

\item{smp_data}{a data frame that needs to comprise all variables named in
\code{fixed} and \code{smp_domains}.}

\item{smp_domains}{a character string containing the name of a variable
that indicates domains in the sample data. The variable can be numeric or a
factor.}

\item{transformation}{a character string. Three different transformation
types for the dependent variable can be chosen (i) no transformation ("no");
(ii) log transformation ("log"); (iii) Box-Cox transformation ("box.cox").}

\item{method}{a character string. In order to determine the optimal parameter
for a Box-Cox transformation six different estimation methods can be chosen
(i) restricted Maximum-Likelihood ("reml"); (ii) pooled skewness
minimization ("p.skew"); (iii) skewness minimization ("skew");
(iv) minimization of Kolmogorov-Smirnoff divergence  ("div.ks");
(v) minimization of Craemer von Mises divergence ("div.cvm"); (vi)
minimization of Kullback Leibner divergence  ("div.kl"). In case of no and
log transformation "NA" can be selected since no optimization is neccessary
for these two transformation types.}

\item{interval}{a numeric vector containing a lower and upper limit
determining an interval for the estimation of the optimal parameter. Defaults
to c(-1,2).}
}
\value{
a scalar returning the optimal lambda.
}
\description{
Function \code{optimal_parameter} estimates the optimal parameter lambda
that is neccessary to determine Box-Cox transformations.
}
\examples{
# Loading data
data("Xoutsamp_AuxVar")
data("incomedata")

opt_par <- optimal_parameter(generic_opt, income~educ1, incomedata,
"provlab", "box.cox", "reml")
}
