% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.compare_trafo.R
\name{plot.trafo_compare}
\alias{plot.trafo_compare}
\title{Plots for linear models with transformed dependent variable}
\usage{
\method{plot}{trafo_compare}(x, ...)
}
\arguments{
\item{x}{an object of type \code{trafo_compare}}

\item{...}{additional arguments that are not used in this method}
}
\description{
For the two transformed models a range of plots is returned in order to check
model assumptions graphically.
}
