% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trafo_mod.R
\name{plot.trafo_mod}
\alias{plot.trafo_mod}
\title{Plot for models with untransformed and transformed dependent variable}
\usage{
\method{plot}{trafo_mod}(x, ...)
}
\arguments{
\item{x}{an object of type \code{trafo_mod}}

\item{...}{additional arguments that are not used in this method}
}
\description{
Plots for comparing the diagnostics of transformed and untransformed models.
}
\keyword{internal}
