% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.transformation.R
\name{plot.transformation}
\alias{plot.transformation}
\title{Plot Transformation}
\usage{
\method{plot}{transformation}(x, ...)
}
\arguments{
\item{x}{an object of type \code{transformation}}

\item{\dots}{additional arguments to be passed to the estimation function; see bcxEst()}
}
\value{
an object of class \code{transformation}; see bcxEst()
}
\description{
Plots object of type \code{transformation}. Four plots are shown: a plot of the profile loglikelihood vs lambda; a qq plot; a histogram of the residuals vs fitted values of the transformed model
}
