% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotit.R
\name{plot_trafolm}
\alias{plot_trafolm}
\title{Plot for optimal transformation parameter - linear models}
\usage{
plot_trafolm(lambdarange, lambdaoptim, measoptim, y, x, trafo, method,
  custom_func, custom_func_std)
}
\arguments{
\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter.}

\item{lambdaoptim}{optimal or given transformation parameter.}

\item{measoptim}{measure at the optimal transformation parameter.}

\item{y}{dependent variable.}

\item{x}{matrix of regressors}

\item{trafo}{a character string that selects the transformation.}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Maximum likelihood approach ("ml"), (ii) Skewness minimization ("skew"),
(iii) Kurtosis optimization ("kurt"), (iv) Divergence minimization by 
Kolmogorov-Smirnov ("div.ks"), by Cramer-von-Mises ("div.cvm") or by 
Kullback-Leibler ("div.kl"). Defaults to "ml". In case of no and
log transformation "NA" can be selected since no optimization is necessary
for these two transformation types.}
}
\description{
Plot for optimal transformation parameter - linear models
}
\keyword{internal}
