% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trafolme.R
\name{plot_trafolme}
\alias{plot_trafolme}
\title{Plot for optimal transformation parameter - linear models}
\usage{
plot_trafolme(lambdarange, lambdaoptim, measoptim, y, formula, data, rand_eff,
  transfor, method)
}
\arguments{
\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter.}

\item{lambdaoptim}{optimal or given transformation parameter.}

\item{measoptim}{measure at the optimal transformation parameter.}

\item{y}{dependent variable.}

\item{formula}{formula object.}

\item{data}{data extracted from the lme object.}

\item{rand_eff}{random effect extracted from lme object.}

\item{transfor}{a character string that selects the transformation.}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Restricted maximum likelihood approach ("reml"), 
(ii) Skewness minimization ("skew") and pooled skewness minimization ("pskew"), 
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cm") or by Kullback-Leibler ("div.kl").}
}
\description{
Plot for optimal transformation parameter - linear models
}
\keyword{internal}
