% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.compare_trafo.R
\name{print.diagnostics.compare_trafo}
\alias{print.diagnostics.compare_trafo}
\title{Prints diagnostics of two trafo objects}
\usage{
\method{print}{diagnostics.compare_trafo}(x, ...)
}
\arguments{
\item{x}{an object of type \code{diagnostics.compare_trafo}}

\item{...}{additional arguments that are not used in this method}
}
\description{
Prints diagnostics of two trafo objects.
}
