% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.trafo_mod.R
\name{print.diagnostics.trafo_lm}
\alias{print.diagnostics.trafo_lm}
\title{Prints diagnostics of an untransformed and a transformed model}
\usage{
\method{print}{diagnostics.trafo_lm}(x, ...)
}
\arguments{
\item{x}{an object of type \code{diagnostics.trafo_lm}}

\item{...}{additional arguments that are not used in this method}
}
\description{
Prints diagnostics of an untransformed and a transformed model.
}
