% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trafo.R
\name{print.summary.trafo}
\alias{print.summary.trafo}
\title{Print summary trafo}
\usage{
\method{print}{summary.trafo}(x, ...)
}
\arguments{
\item{x}{an object of type \code{trafo}}

\item{...}{additional arguments that are not used in this method}
}
\description{
prints objects to be shown in the summary function for objects of type 
\code{transformation}
}
