% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.compare_trafo.R
\name{print.summary.trafo_compare}
\alias{print.summary.trafo_compare}
\title{Prints summary of trafo_compare objects}
\usage{
\method{print}{summary.trafo_compare}(x, ...)
}
\arguments{
\item{x}{an object of type \code{summary.trafo_compare}}

\item{...}{additional arguments that are not used in this method}
}
\description{
Prints objects to be shown in the summary function for objects of 
type \code{trafo_compare}.
}
