% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trafo_mod.R
\name{print.summary.trafo_lm}
\alias{print.summary.trafo_lm}
\title{Print summary trafo}
\usage{
\method{print}{summary.trafo_lm}(x, ...)
}
\arguments{
\item{x}{an object of type \code{summary.trafo_lm}}

\item{...}{additional arguments that are not used in this method}
}
\description{
prints objects to be shown in the summary function for objects of 
type \code{trafo_lm}
}
