% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trafo_mod.R
\name{print.summary.trafo_mod}
\alias{print.summary.trafo_mod}
\title{Print summary trafo}
\usage{
\method{print}{summary.trafo_mod}(x, ...)
}
\arguments{
\item{x}{an object of type \code{summary.trafo_mod}}
}
\description{
prints objects to be shown in the summary function for objects of 
type \code{trafo_mod}
}
\keyword{internal}
