% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{print.summary.transformation}
\alias{print.summary.transformation}
\title{Print summary transformations}
\usage{
\method{print}{summary.transformation}(x, ...)
}
\arguments{
\item{x}{an object of type \code{transformation}}
}
\description{
prints objects to be shown in the summary function for objects of type \code{transformation}
}
\keyword{internal}
