% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trafo.R
\name{print.trafo}
\alias{print.trafo}
\title{Prints object of type trafo}
\usage{
\method{print}{trafo}(x, ...)
}
\arguments{
\item{x}{an object of type \code{trafo}.}

\item{...}{other parameters that can be passed to the function.}
}
\description{
Prints object of type trafo
}
