% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.compare_trafo.R
\name{print.trafo_compare}
\alias{print.trafo_compare}
\title{Prints object of type trafo_compare}
\usage{
\method{print}{trafo_compare}(x, ...)
}
\arguments{
\item{x}{an object of type trafo_compare.}

\item{...}{other parameters that can be passed to the function.}
}
\description{
Prints object of type trafo_compare
}
