% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trafo_mod.R
\name{print.trafo_lm}
\alias{print.trafo_lm}
\title{Prints object of type trafo_lm}
\usage{
\method{print}{trafo_lm}(x, ...)
}
\arguments{
\item{x}{an object of type \code{trafo_lm}.}

\item{...}{other parameters that can be passed to the function.}
}
\description{
Prints object of type trafo_lm
}
