% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trafo_mod.R
\name{print.trafo_mod}
\alias{print.trafo_mod}
\title{Print object of type transformation}
\usage{
\method{print}{trafo_mod}(x, ...)
}
\arguments{
\item{x}{an object of type trafo.}

\item{...}{other parameters that can be passed to the function.
#@keywords internal}
}
\description{
Print object of type transformation
}
