% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglook_methods.R
\name{reglook}
\alias{reglook}
\alias{reglook.default}
\alias{reglook.data.frame}
\title{Search an object by regular expression.}
\usage{
reglook(object, ...)

\method{reglook}{default}(object, regex, ...)

\method{reglook}{data.frame}(object, regex, keys = NULL, multiple = c("OR", "AND"))
}
\arguments{
\item{object}{an object to be searched.}

\item{...}{extra arguments, currently none.}

\item{regex}{a regular expression.}

\item{keys}{a vector of data frame variable names to be included
in the search. Defaults to NULL, which means that all variables will be
looked up.}

\item{multiple}{ignored when a single search key is provided.
It describes how searching results from multiple data frame columns will
be handled. 'OR' (default) specifies that the multiply column search results
will be merged by logical sum, for 'AND' they will be merged by logical
intersection.}
}
\value{
an object of the same class as the input one.
}
\description{
Select elements of a vector or data frame records which match
the given regular expression.
}
