% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_transformation.R
\name{safely_mutate}
\alias{safely_mutate}
\alias{safely_select}
\title{Select and manipulate data frame columns in an error-resistant mode.}
\usage{
safely_mutate(.data, ...)

safely_select(.data, cols, fill = NULL)
}
\arguments{
\item{.data}{a data frame.}

\item{...}{expressions for modification of data frame columns, see:
\code{\link[dplyr]{mutate}} for details.}

\item{cols}{a character vector with names of columns to select.}

\item{fill}{value to be inputed. If not `NULL`, any non-existing columns
will be created and filled with this value.}
}
\value{
a data frame.
}
\description{
The functions select and transform columns of a data frame in an analogical
way as \code{\link[dplyr]{select}} and \code{\link[dplyr]{mutate}} but in an
error-resistant mode. This means that e.g. modification of
non-existing columns does not throw an error. In case of any errors, the
unmodified input data frame is returned (`safely_mutate()`) or only the
columns found in the data frame are selected (`safely_select()`).
}
