% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTransformation.default.R
\name{selectTransformation.default}
\alias{selectTransformation.default}
\title{Transforma Default}
\usage{
\method{selectTransformation}{default}(object, ...)
}
\arguments{
\item{object}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{\dots}{additional arguments to be passed to the estimation function; see doselectTransformation()}
}
\value{
an object of class \code{transformation}; see doselectTransformation()
}
\description{
Chooses the 'best' transformation in the sense of AIC, R.Squared, Skewness or Kurtosis; For positive y compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum
}
\keyword{internal}
