% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTransformation.formula.R
\name{selectTransformation.formula}
\alias{selectTransformation.formula}
\title{Select transformartion formula}
\usage{
\method{selectTransformation}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{object of type formula with the model to transform}

\item{data}{an optional data frame containing the variables in the model}

\item{\dots}{additional arguments to be passed to the estimation function;  see doselectTransformation()}
}
\value{
an object of class \code{transformation}; see doselectTransformation()
}
\description{
Chooses the 'best' transformation in the sense of AIC, R.Squared, Skewness or Kurtosis; For positive y compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum
}
\keyword{internal}

