% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTransformation.lm.R
\name{selectTransformation.lm}
\alias{selectTransformation.lm}
\title{Transforma lm}
\usage{
\method{selectTransformation}{lm}(object, ...)
}
\arguments{
\item{object}{of type lm with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see doselectTransformation()}
}
\value{
an object of class \code{transformation}; see doselectTransformation()
}
\description{
Chooses the 'best' transformation in the sense of AIC, R.Squared, Skewness or Kurtosis; For positive y compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum
}
\keyword{internal}

