% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTransformation.R
\name{selectTransformation}
\alias{selectTransformation}
\title{Selects Transformation}
\usage{
selectTransformation(object, ...)
}
\arguments{
\item{object}{a data.frame, matrix, formula or fitted model; see below.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
An object of class \code{transformation} with the following arguments

llike The value of profile log-likelihood at its maximum

logvector The profile log-likelihood evaluated at \code{lambdavector}

lambdavector Employed family of transformations

A sequence with optional values for \code{lambda}

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Chooses the 'best' transformation in the sense of AIC, R.Squared, Skewness or Kurtosis; For positive \code{y} compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum.
}
\keyword{selecTransformation}
