% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_colnames}
\alias{set_colnames}
\title{Set column names of a data frame or matrix.}
\usage{
set_colnames(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Sets column names of a matrix, data frame or tibble.
}
