% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_rownames}
\alias{set_rownames}
\title{Set rownames of a data frame.}
\usage{
set_rownames(x, row_names)
}
\arguments{
\item{x}{a data frame.}

\item{rown_names}{a vector of unique rownames.}
}
\value{
a data frame.
}
\description{
Sets rownames of a data frame. A tibble is sielnty connverted
to a classical data frame.
}
