% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estims.R
\name{skewness_min}
\alias{skewness_min}
\title{Skewness minimization by Molina}
\usage{
skewness_min(res = res)
}
\arguments{
\item{res}{residuals from a linear model with response variable yt and 
explanatory variables x}
}
\value{
absolute value of the skewness of the residuals
}
\description{
Skewness minimization by Molina
}
\keyword{internal}
