% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtshift.R
\name{sqrtshift}
\alias{sqrtshift}
\title{sqrtshift Transformation}
\usage{
sqrtshift(object, ...)
}
\arguments{
\item{object}{an object of type lm or lme.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
The return depends on the class of its argument. The 
documentation of particular methods gives detailed information about the 
return of that method.
}
\description{
Transforms the dependent variable of a regression model using the sqrtshift
transformation. The regression model can be a linear or a linear mixed model 
with one random intercept.
}
\seealso{
\code{\link{sqrtshift.lm}},  \code{\link{sqrtshift.lme}}
}
\keyword{Sqrt}
\keyword{shift}
