% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.compare_trafo.R
\name{summary.compare_trafo}
\alias{summary.compare_trafo}
\title{Summary for linear models with transformed dependent variable}
\usage{
\method{summary}{compare_trafo}(object, ...)
}
\arguments{
\item{object}{an object of type \code{compare_trafo}}

\item{...}{additional arguments that are not used in this method}
}
\value{
an object of class \code{summary.compare_trafo}
}
\description{
The summary contains the summary for both transformed models. The summary is 
based on the summary for objects of type \code{lm}. The returned object is 
suitable for printing with the print.summary.compare_trafo method.
}
