% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trafo.R
\name{summary.trafo}
\alias{summary.trafo}
\title{Summary of trafo objects}
\usage{
\method{summary}{trafo}(object, ...)
}
\arguments{
\item{object}{an object of type \code{trafo}}

\item{...}{additional arguments that are not used in this method}
}
\value{
an object of class \code{summary.trafo}
}
\description{
Information about the residuals from a transformed model. The returned object 
is suitable for printing with the print.summary.trafo method.
}
\examples{
# Load data
data("eusilcA_Vienna")

# Fit linear model
lm_Vienna <- lm(eqIncome ~ eqsize + gender + cash + unempl_ben + age_ben + 
rent + cap_inv + tax_adj + dis_ben + sick_ben + surv_ben + 
fam_allow + house_allow, data = eusilcA_Vienna)

# Transform dependent variable using skewness minimization
boxcox_trafo <- boxcox(object = lm_Vienna, lambda = "estim", method = "skew",
plotit = FALSE)

# Get plots
summary(boxcox_trafo)
}
