% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.compare_trafo.R
\name{summary.trafo_compare}
\alias{summary.trafo_compare}
\title{Summary for two differently transformed models}
\usage{
\method{summary}{trafo_compare}(object, ...)
}
\arguments{
\item{object}{an object of type \code{trafo_compare}}

\item{...}{additional arguments that are not used in this method}
}
\value{
An object of class \code{summary.trafo_compare}. The method 
\code{\link{print.summary.trafo_compare}} can be used for this class.
}
\description{
The summary contains the summary for two transformed models. The summary is 
based on the summary for objects of type \code{lm}.
}
