% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trafo_mod.R
\name{summary.trafo_mod}
\alias{summary.trafo_mod}
\title{Summary for linear models with untransformed and transformed dependent 
variable}
\usage{
\method{summary}{trafo_mod}(object, ...)
}
\arguments{
\item{object}{an object of type \code{trafo_mod}}

\item{...}{additional arguments that are not used in this method}
}
\value{
An object of class \code{summary.trafo_mod}. The method 
\code{\link{print.summary.trafo_mod}} can be used for this class.
}
\description{
The summary method for class \code{trafo_mod} contains a summary for an 
untransformed and a transformed model. The resulting summary is based on the 
summary for objects of type \code{lm}.
}
