% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.transformation.R
\name{summary.transformation}
\alias{summary.transformation}
\title{Summary transformations}
\usage{
\method{summary}{transformation}(object, ...)
}
\arguments{
\item{...}{additional arguments that are not used in this method}

\item{x}{an object of type \code{transformation}}
}
\value{
an object of class \code{summary.transformation}
}
\description{
Information about the transformed data and model and components of an 
transformation object are extracted. The returned object is suitable for 
printing with the print.summary.transformation method.
}
\keyword{internal}
