% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/arrangement_methods.R
\name{to_head}
\alias{to_head}
\alias{to_tail}
\alias{to_head.default}
\alias{to_tail.default}
\alias{to_head.matrix}
\alias{to_tail.matrix}
\alias{to_head.data.frame}
\alias{to_tail.data.frame}
\alias{to_head.tbl_df}
\alias{to_tail.tbl_df}
\title{Shift elements or observations to the object's head or tail.}
\usage{
to_head(x, ...)

to_tail(x, ...)

\method{to_head}{default}(x, idx, ...)

\method{to_tail}{default}(x, idx, ...)

\method{to_head}{matrix}(x, idx, ...)

\method{to_tail}{matrix}(x, idx, ...)

\method{to_head}{data.frame}(x, idx, ...)

\method{to_tail}{data.frame}(x, idx, ...)

\method{to_head}{tbl_df}(x, idx, ...)

\method{to_tail}{tbl_df}(x, idx, ...)
}
\arguments{
\item{x}{an object: an atomic vector, list, matrix, or data frame.}

\item{...}{extra arguments passed to the methods.}

\item{idx}{a vector of indexes or names of objects to be placed at the object's
start or end.}
}
\value{
an object of the same class as `x` with the selected elements or
rows placed at it's start or end.
}
\description{
Function `to_head()` shifts selected elements of an atomic vector or a list,
or selected rows of a matrix or a data frame to the object's head, i.e.
starting position.
Function `to_tail()` places the selected elements or observations at the
object's tails, i.e. end position.
}
\details{
The elements r observations may be specified by their positions or a names.
The order of the shifted elements corresponds to the position/name vector.
If the object has no names, an error is raised.
}
