% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo_back.R
\name{trafo_back}
\alias{trafo_back}
\title{Back-transforms vectors with used transformation in trafo_lm}
\usage{
trafo_back(object, prediction)
}
\arguments{
\item{object}{an object of type \code{trafo_lm}.}

\item{prediction}{the return of the predict.lm method.}
}
\value{
The backtransformed prediction.
}
\description{
The function transforms vectors as the prediction or confidence intervals 
back to the original scale of the used transformation in trafo_lm.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Compare untransformed and transformed model
trafo_cars <- trafo_lm(object = lm_cars, trafo = "bickeldoksum", method = "skew", 
lambdarange = c(1e-11, 2))

# Back-transform prediction and confidence interval
trafo_back(trafo_cars, predict(trafo_cars$trafo_mod, interval = "confidence"))

}
