% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo_lme.R
\name{trafo_lme}
\alias{trafo_lme}
\title{Fits transformed linear mixed models}
\usage{
trafo_lme(object, trafo = "boxcox", lambda = "estim", method = "reml",
  lambdarange = c(-2, 2), std = FALSE, custom_trafo = NULL)
}
\arguments{
\item{object}{an object of type lm or lme with the model to transform}

\item{trafo}{a character string. Different transformations can be used.}

\item{lambda}{either a character named "estim" if the optimal transformation
parameter should be estimated or a numeric value determining a given 
transformation parameter.}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Restricted maximum likelihood approach ("reml"), 
(ii) Skewness minimization ("skew") and pooled skewness minimization ("pskew"), 
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cvm") or by Kullback-Leibler ("div.kl").}

\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter.}

\item{std}{logical. If TRUE, the transformed model is returned based on the 
standardized transformation.}

\item{custom_trafo}{a list that determines a one parameter transformation and
the standardized one parameter transformation.}
}
\value{
an object of class \code{trafo_mod}.
}
\description{
Function \code{trans_lme} fits linear mixed models with one random effect 
and transformed dependent variable. The return are two lme objects where the 
first is the transformed and the second the untransformed linear mixed model.
}
\keyword{internal}
