% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo_predict.R
\name{trafo_predict}
\alias{trafo_predict}
\title{Predict method for linear models with transformed dependent variable}
\usage{
trafo_predict(object, newdata = NULL, interval = "none",
  level = 0.95)
}
\arguments{
\item{object}{an object of type \code{trafo_lm}.}

\item{newdata}{an optional data frame in which to look for variables with 
which to predict. If omitted, the fitted values are used.}

\item{interval}{type of interval calculation. Possible values: "none", "confidence",
"prediction". Defaults to "none".}

\item{level}{tolerance/confidence level.}
}
\value{
A vector of predictions or a matrix of predictions and bounds with 
column names fit, lwr and upr if interval is set.
}
\description{
The function returns predicted values based on the linear model. The 
predicted values are back-transformed corresponding to the transformation 
used in the model. Note that the back-transformation can induce a bias.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Compare untransformed and transformed model
trafo_cars <- trafo_lm(object = lm_cars, trafo = "bickeldoksum", method = "skew", 
lambdarange = c(1e-11, 2))

# Get predictions in the back-transformed scale
trafo_predict(trafo_cars)
}
