% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforma.R
\docType{package}
\name{transforma}
\alias{transforma}
\alias{transforma-package}
\title{transforma: A Package for compute optimal parameters for different families of transformations}
\description{
The transforma package uses the maximum likelihood approach to compute
}
\section{The transforma functions are}{
NA
}

