\name{transformar.lm}
\alias{transformar.lm}
\title{Select a transform lm}
\usage{
\method{transformar}{lm}(object)
}
\arguments{
\item{lm}{object of type lm with the model to transform}
}
\value{
Optimal value for \code{lambda} and value of \code{profile log-likelihood}
}
\description{
Select a transform
}
\keyword{internal}

