% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woparam.lm.R
\name{woparam.lm}
\alias{woparam.lm}
\title{Transformations without parameter for linear models}
\usage{
\method{woparam}{lm}(object, trafo, custom_trafo = NULL, ...)
}
\arguments{
\item{object}{an object of type lm.}

\item{trafo}{character that determines the transformation.}

\item{custom_trafo}{a function that specifies a transformation without 
parameter that needs to be estimated or given.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
an object of class \code{trafo}.
}
\description{
The function transforms the dependent variable of a linear model using 
different transformations without parameter.
}
\keyword{internal}
