% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woparam.R
\name{woparam}
\alias{woparam}
\title{Transformations without transformation parameter}
\usage{
woparam(object, ...)
}
\arguments{
\item{object}{an object of type lm or lme.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
The return depends on the class of its argument. The 
documentation of particular methods gives detailed information about the 
return of that method.
}
\description{
Transforms the dependent variable of a regression model using
transformations without transformation parameter.
}
\seealso{
\code{\link{woparam.lm}}
}
\keyword{internal}
