% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YeoJohnson.default.R
\name{yeojohnson.default}
\alias{yeojohnson.default}
\title{Yeo-Johnson Default}
\usage{
\method{yeojohnson}{default}(object, ...)
}
\arguments{
\item{object}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{\dots}{additional arguments to be passed to the estimation function;  see yeojohnsonEst()}
}
\value{
an object of class \code{transformation}; see yeojohnsonEst()
}
\description{
Yeo-Johnson estimation
}
\keyword{internal}

