% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YeoJohnson.formula.R
\name{yeojohnson.formula}
\alias{yeojohnson.formula}
\title{Yeo-Johnson Formula}
\usage{
\method{yeojohnson}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{object of type formula with the model to transform}

\item{data}{an optional data frame containing the variables in the model}

\item{\dots}{additional arguments to be passed to the estimation function;  see yeohjohnsonEst()}
}
\value{
an object of class \code{transformation}; see yeohjohnsonEst()
}
\description{
Yeo-Johnson estimation
}
\keyword{internal}

