% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeojohnson.lme.R
\name{yeojohnson.lme}
\alias{yeojohnson.lme}
\title{Yeo-Johnson lme}
\usage{
\method{yeojohnson}{lme}(object, method, lambdarange = c(-2, 2),
  tol = 1e-04, ...)
}
\arguments{
\item{object}{of type lme with the model to transform}

\item{\dots}{additional arguments to be passed to the estimation function; see yeojohnsonEst()}
}
\value{
an object of class \code{transformation}; see yeojohnsonEst()
}
\description{
Yeo-Johnson estimation
}
\keyword{internal}
