% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeojohnson.R
\name{yeojohnson}
\alias{yeojohnson}
\title{Yeo-Johnson Transformation}
\usage{
yeojohnson(object, ...)
}
\arguments{
\item{object}{an object of type lm or lme.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
The return depends on the class of its argument. The 
documentation of particular methods gives detailed information about the 
return of that method.
}
\description{
Transforms the dependent variable of a regression model using the Yeo-Johnson
transformation. The regression model can be a linear or a linear mixed model 
with one random intercept.
}
\seealso{
\code{\link{yeojohnson.lm}},  \code{\link{yeojohnson.lme}}
}
\keyword{Yeo-Johnson}
