% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trafo.R
\name{as.data.frame.trafo}
\alias{as.data.frame.trafo}
\title{Data frame with transformed variables}
\usage{
\method{as.data.frame}{trafo}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an object of type \code{trafo}.}

\item{row.names}{NULL or a character vector giving the row names for the 
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column 
names (to syntactic names: see make.names) is optional. Note that all of R's 
base package as.data.frame() methods use optional only for column names 
treatment, basically with the meaning of 
data.frame(*, check.names = !optional)}

\item{...}{other parameters that can be passed to the function.}
}
\description{
The data frame that is returned contains the variables that are used
in the model and additionally a variable with the transformed dependent
variable. To the variable name of the dependent variable a t is added for
transformed.
}
