% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions}
\alias{assumptions}
\title{First check of assumptions}
\usage{
assumptions(object, method = "ml")
}
\arguments{
\item{object}{an object of type \code{lm}.}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Maximum likelihood approach ("ml"), (ii) Skewness minimization ("skew"),  
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cvm") or by Kullback-Leibler ("div.kl"). Defaults
to "ml".}
}
\value{
an object of class \code{diagnostics.compare_trafo}
}
\description{
Gives a first overview if a transformation is useful and which transformation
is promising to fulfill the model assumptions normality, homoscedasticity and 
linearity.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Transform with Bickel-Doksum transformation
bd_trafo <- bickeldoksum(object = lm_cars, plotit = FALSE)

# Transform with Box-Cox transformation
bc_trafo <- boxcox(object = lm_cars, method = "skew", plotit = FALSE)

# Compare transformed models
compare <- compare_trafo(object = lm_cars, trafos = list(bd_trafo, bc_trafo), 
std = FALSE)

# Get diagnostics
diagnostics(compare)
}
