% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{Box-Cox Transformation}
\usage{
boxcox(object, ...)
}
\arguments{
\item{object}{a data.frame, matrix, formula or fitted model; see below.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
An object of class \code{transformation} with the following arguments

llike The value of profile log-likelihood at its maximum

logvector The profile log-likelihood evaluated at \code{lambdavector}

lambdavector A sequence with optional values for \code{lambda}

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Estimates the transformation parameter for the Box-Cox transformation employing the profile log-likehood. Only accepts positive values for the dependent variable.
}
\keyword{boxcox}
