% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_cx.R
\name{bx_cx}
\alias{bx_cx}
\title{Box-Cox Transformation}
\usage{
bx_cx(object, ...)
}
\arguments{
\item{object}{a data.frame, matrix, formula or fitted model; see below.}

\item{data}{an optional data frame, list or environment as in lm.}

\item{XY}{a data.frame or a matrix, whose first column is the dependent variable and the rest of the columns are the independent variables.}

\item{lambdarange}{range for the estimation parameter lambda - c(-2,2)}
}
\value{
An object of class \code{transformation} with the following arguments

llike The value of profile log-likelihood at its maximum

logvector The profile log-likelihood evaluated at \code{lambdavector}

lambdavector A sequence with optional values for \code{lambda}

family Employed family of transformations

yt Vector of the transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable
}
\description{
Estimates the transformation parameter for the Box-Cox transformation employing the profile log-likehood. Only accepts positive values for the dependent variable.
}
\keyword{boxcox}

