% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trafo.R
\name{compare_trafo}
\alias{compare_trafo}
\title{Compares transformed linear models}
\usage{
compare_trafo(object, trafos, std)
}
\arguments{
\item{object}{an object of type lm}

\item{trafos}{a list of \code{trafo} objects based on the same model given 
in object.}

\item{std}{logical. If TRUE, the transformed models are returned based on the 
standardized transformation.}
}
\value{
an object of class \code{compare_trafo}.
}
\description{
Function \code{compare_trafo} compares linear models with transformed 
dependent variable.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Transform with Bickel-Doksum transformation
bd_trafo <- bickeldoksum(object = lm_cars, plotit = FALSE)

# Transform with Box-Cox transformation
bc_trafo <- boxcox(object = lm_cars, method = "skew", plotit = FALSE)

# Compare transformed models
compare_trafo(object = lm_cars, trafos = list(bd_trafo, bc_trafo), 
std = FALSE)
}
