% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trafo.R
\name{compare_trafo}
\alias{compare_trafo}
\title{Compares transformed linear models}
\usage{
compare_trafo(object, trafos, std)
}
\arguments{
\item{object}{an object of type lm or lme with the model to transform}

\item{trafos}{a list of trafo objects based on the same model as given 
in object.}

\item{std}{logical. If TRUE, the transformed model is returned based on the 
standardized transformation.}
}
\value{
an object of class \code{compare_trafo}.
}
\description{
Function \code{compare_trafo} fits linear models with transformed dependent 
variable.
}
