% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTransformation.R
\name{compareTransformation}
\alias{compareTransformation}
\title{Compares Transformations}
\usage{
compareTransformation(object, ...)
}
\arguments{
\item{object}{a data.frame, matrix, formula or fitted model; see below.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
A data frame comparing families of transformations. If plotit ==  TRUE, plots comparing transformations are shown as well.
}
\description{
Compares families of transformations with respect to  sense of AIC, R.Squared, Skewness or Kurtosis; For positive y compares between Box-Cox, Dual and Manly transformations; For all values of y it compares between Yeo-Johnson, Manly, Modulus and Bickel-Doksum.
}
