% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_methods.R
\name{compress.default}
\alias{compress.default}
\title{Convert a named object into a tibble.}
\usage{
\method{compress}{default}(x, names_to = "names", values_to = "values", ...)
}
\arguments{
\item{x}{a named object, e.g. atomic vector, to be converted to a tibble.}

\item{names_to}{name of the variable storing the names.}

\item{values_to}{name of the variable storing the object's elements.}

\item{...}{extra arguments, currently none.}
}
\value{
a tibble.
}
\description{
Converts a named object, e.g. an atomic vector into a tibble.
The object elements are stored in a variable named as specified by the
'values_to' argument, the names are stored under 'names_to' variable.
}
