% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.trafo_mod.R
\name{diagnostics.trafo_mod}
\alias{diagnostics.trafo_mod}
\title{Diagnostics for models with untransformed and transformed dependent variable}
\usage{
\method{diagnostics}{trafo_mod}(object, ...)
}
\arguments{
\item{object}{an object of type \code{compare_trafo}}

\item{...}{additional arguments that are not used in this method}
}
\value{
an object of class \code{summary.transformation}
}
\description{
Information about the transformed data and model and components of an 
transformation object are extracted. The returned object is suitable for 
printing with the print.summary.transformation method.
}
