% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_lm.R
\name{estim_lm}
\alias{estim_lm}
\title{Wrapper function for estimation methods - linear models}
\usage{
estim_lm(lambda, y, x, method, transfor)
}
\arguments{
\item{lambda}{transformation parameter}

\item{y}{vector of response variables}

\item{x}{matrix of regressors}

\item{method}{a character string. In order to determine the optimal parameter
for the transformation five different estimation methods can be chosen
(i) Maximum-Likelihood ("ml"); (ii) skewness minimization ("skew");
(iii) minimization of Kolmogorov-Smirnoff divergence  ("div.ks");
(iv) minimization of Craemer von Mises divergence ("div.cvm"); (v)
minimization of Kullback Leibner divergence  ("div.kl"). In case of no and
log transformation "NA" can be selected since no optimization is neccessary
for these two transformation types.}
}
\value{
Depending on the selected \code{method} the return is a log
likelihood, a skewness, a pooled skewness or a Kolmogorov-Smirnoff, Craemer
von Mises or Kullback Leibner divergence.
}
\description{
Wrapper function for estimation methods - linear models
}
\keyword{internal}
