% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_lme.R
\name{estim_lme}
\alias{estim_lme}
\title{Wrapper function for estimation methods - linear mixed models}
\usage{
estim_lme(lambda, y, formula, data, rand_eff, method, trafo)
}
\arguments{
\item{lambda}{transformation parameter}

\item{y}{vector of response variables}

\item{formula}{a formula object that contains the dependent and the explanatory
measures}

\item{data}{the data.frame that is given to functio nlme and that contains the 
regression variables.}

\item{rand_eff}{the random effect extracted from the lme object.}

\item{method}{a character string. In order to determine the optimal parameter
for the transformation five different estimation methods can be chosen
(i) Maximum-Likelihood ("ml"); (ii) skewness minimization ("skew");
(iii) minimization of Kolmogorov-Smirnoff divergence  ("div.ks");
(iv) minimization of Craemer von Mises divergence ("div.cvm"); (v)
minimization of Kullback Leibner divergence  ("div.kl"). In case of no and
log transformation "NA" can be selected since no optimization is neccessary
for these two transformation types.}

\item{trafo}{a character string that selects the transformation.}
}
\value{
Depending on the selected \code{method} the return is a log
likelihood, a skewness, a pooled skewness or a Kolmogorov-Smirnoff, Craemer
von Mises or Kullback Leibner divergence.
}
\description{
Wrapper function for estimation methods - linear mixed models
}
\keyword{internal}
