% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_methods.R
\name{exchange.default}
\alias{exchange.default}
\title{Substitute/exchange object elements.}
\usage{
\method{exchange}{default}(x, dict, key = "variable", value = "label", ...)
}
\arguments{
\item{x}{an object, e.g an atomic vector or a matrix.}

\item{dict}{a data frame with the recoding scheme. It has to contain columns
specified by the key and value arguments.}

\item{key}{the name of the dict data frame variable corresponding to
the elements of x: the search key.}

\item{value}{the name of the dict data frame variable containing the elements
used for substitution of x: the returned labels/values.}

\item{...}{extra arguments, currently none.}
}
\value{
a vector.
}
\description{
Substitutes elements of a vector or other object with values
provided by a dictionary: a data frame with the re-coding scheme.
}
\details{
Any numeric object is silently coerced
to a character vector.
}
