% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{full_rbind}
\alias{full_rbind}
\title{Full row binding.}
\usage{
full_rbind(x, y, ...)
}
\arguments{
\item{x}{an object.}

\item{y}{an object.}

\item{...}{additional arguments passed to methods.}
}
\value{
a data frame, a tibble or matrix.
}
\description{
Binds two data frames, matrices or vectors by rows,
missing variables are filled with NA
}
