% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{logtrafo}
\alias{logtrafo}
\title{Log transformation for linear models}
\usage{
logtrafo(object)
}
\arguments{
\item{object}{an object of type lm.}
}
\value{
an object of class \code{trafo}.
}
\description{
The function transforms the dependent variable of a linear model using the 
Log transformation. The transformation parameter can either be 
estimated using different estimation methods or given.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Transform dependent variable 
logtrafo(object = lm_cars)
}
