% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manly.lm.R
\name{manly.lm}
\alias{manly.lm}
\title{Manly lme}
\usage{
\method{manly}{lm}(object, method, lambdarange = c(-2, 2), tol = 1e-04, ...)
}
\arguments{
\item{\dots}{additional arguments to be passed to the estimation function; see manlyEst()}

\item{lm}{object of type lm with the model to transform}
}
\value{
an object of class \code{transformation}; see manlyEst()
}
\description{
Manly estimation
}
\keyword{internal}
