% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manly.lme.R
\name{manly.lme}
\alias{manly.lme}
\title{Manly transformation for linear and linear mixed models}
\usage{
\method{manly}{lme}(object, lambda, method, lambdarange = c(-2, 2),
  plotit = TRUE, ...)
}
\arguments{
\item{object}{an object of type lm or lme with the model to transform}

\item{lambda}{either a character named "estim" if the optimal lambda should
be estimated or a numeric value determining a given lambda.}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter. 
(i) Maximum likelihood approaches: for linear models maximum likelihood ("ML")
and for linear mixed models restricted maximum likelihood ("reml"); 
(ii) Skewness minimizations: for linear models only skewness minimization 
("skew") and for linear mixed models also pooled skewness minimization; 
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cm") or by Kullback-Leibler ("div.kl") for both 
model types.}

\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter. 
Defaults to \code{c(-2, 2)} for the Manly transformation.}

\item{plotit}{logical. If TRUE, a plot that illustrates the optimal 
transformation parameter is returned.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
an object of class \code{transformation}
}
\description{
Depending on the class of the first object, this function estimates the 
optimal transformation parameter for the Manly transformation for 
the model given to the function.
}
\keyword{internal}
