% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manly.R
\name{manly}
\alias{manly}
\title{Manly Transformation}
\usage{
manly(object, ...)
}
\arguments{
\item{object}{a data.frame, matrix, formula or fitted model; see below.}

\item{...}{other parameters that can be passed to the function.}
}
\value{
An object of class \code{transformation} with the following arguments.

llike The value of profile log-likelihood at its maximum.

logvector The profile log-likelihood evaluated at \code{lambdavector}.

lambdavector A sequence with optional values for \code{lambda}.

family Employed family of transformations.

yt Vector of the transformed response variable \code{y}.

modelt An object of type \code{lm} employing the transformed vector \code{yt} as the response variable.
}
\description{
Estimates the transformation parameter for the Manly transformation employing the profile log-likehood.
}
\keyword{Manly}
