% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{modulus}
\alias{modulus}
\title{Modulus Transformation}
\usage{
modulus(object, ...)
}
\arguments{
\item{object}{ formula or fitted model; see below.}

\item{data}{an optional data frame, list or environment as in lm.}

\item{lambdarange}{range for the estimation parameter lambda - default c(-2, 2)}
}
\value{
An object of class \code{transformation} with the following arguments.

llike The value of profile log-likelihood at its maximum.

lambdahat The value of lambda for which the log-likelihood is maximized

logvector The profile log-likelihood evaluated at \code{lambdavector}.

lambdavector A sequence with optional values for \code{lambda}.

family Employed family of transformations.

yt Vector of the transformed response variable \code{y}.

zt Vector of the (normalized) transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{zt} as the response variable.
}
\description{
Estimates the transformation parameter for the Modulus transformation employing the profile log-likehood.
}
\references{
John J. A., and Draper, N. R. . 1980 An Alternative Family of Transformations. \emph{Journal of the Royal Statistical Society}, \bold{29(2)}, 190--197.
}

\keyword{modulus}
