% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{modulus}
\alias{modulus}
\title{Modulus transformation for linear models}
\usage{
modulus(object, lambda = "estim", method = "ml", lambdarange = c(-2, 2),
  plotit = TRUE)
}
\arguments{
\item{object}{an object of type lm.}

\item{lambda}{either a character named "estim" if the optimal transformation
parameter should be estimated or a numeric value determining a given 
transformation parameter. Defaults to "estim".}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Maximum likelihood approach ("ml"), (ii) Skewness minimization ("skew"),  
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cvm") or by Kullback-Leibler ("div.kl"). Defaults to
"ml".}

\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter. 
Defaults to \code{c(-2, 2)}.}

\item{plotit}{logical. If TRUE, a plot that illustrates the optimal 
transformation parameter or the given transformation parameter is returned.}
}
\value{
an object of class \code{trafo}.
}
\description{
The function transforms the dependent variable of a linear model using the 
Modulus transformation. The transformation parameter can either be 
estimated using different estimation methods or given.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Transform dependent variable with fixed lambda 
modulus(object = lm_cars, lambda = 0.8, plotit = FALSE)
}
\references{
John JA, Draper NR (1980). An alternative family of transformations. Journal 
of the Royal Statistical Society: Series C, Vol. 29, 190-197.
}
