% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trafo_mod.R
\name{plot.trafo_lm}
\alias{plot.trafo_lm}
\title{Plot for regression models with untransformed and transformed dependent 
variable}
\usage{
\method{plot}{trafo_lm}(x, which = "all", ...)
}
\arguments{
\item{x}{an object of type \code{trafo_lm}}

\item{which}{one element character that determines the return of plots. For 
single plot returns the possible values are: "qqplot", "hist", "residVsFitted", 
"residVsObs", "scatter", "cooks", "scaleLoc", "residLev". The default is set 
to "all".}

\item{...}{additional arguments that are not used in this method}
}
\description{
For the untransformed and transformed model a range of plots is returned in 
order to check model assumptions graphically.
}
