% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.transformation.R
\name{plot.transformation}
\alias{plot.transformation}
\title{Plot Transformation}
\usage{
plot.transformation(ans)
}
\arguments{


\item{ans}{object of type \code{transformation}}
}

\description{
Plots object of type \code{transformation}. Four plots are shown: a plot of the profile loglikelihood vs lambda; a qq plot; a histogram of the residuals vs fitted values of the transformed model
}
