% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trafolm.R
\name{plot_trafolm}
\alias{plot_trafolm}
\title{Plot for optimal transformation parameter - linear models}
\usage{
plot_trafolm(lambdarange, lambdaoptim, measoptim, y, x, transfor, method)
}
\arguments{
\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter.}

\item{lambdaoptim}{optimal or given transformation parameter.}

\item{measoptim}{measure at the optimal transformation parameter.}

\item{y}{dependent variable.}

\item{x}{matrix of regressors}

\item{transfor}{a character string that selects the transformation.}

\item{method}{a character string. In order to determine the optimal parameter
for the transformation five different estimation methods can be chosen
(i) Maximum-Likelihood ("ml"); (ii) skewness minimization ("skew");
(iii) minimization of Kolmogorov-Smirnoff divergence  ("div.ks");
(iv) minimization of Craemer von Mises divergence ("div.cvm"); (v)
minimization of Kullback Leibner divergence  ("div.kl"). In case of no and
log transformation "NA" can be selected since no optimization is neccessary
for these two transformation types.}
}
\description{
Plot for optimal transformation parameter - linear models
}
\keyword{internal}
