% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.compare_trafo.R
\name{print.compare_trafo}
\alias{print.compare_trafo}
\title{Print object of type compare_trafo}
\usage{
\method{print}{compare_trafo}(x, ...)
}
\arguments{
\item{x}{an object of type compare_trafo.}

\item{...}{other parameters that can be passed to the function.}
}
\description{
Print object of type compare_trafo
}
