% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.compare_trafo.R
\name{print.diagnostics.compare_trafo}
\alias{print.diagnostics.compare_trafo}
\title{Print diagnostics trafo}
\usage{
\method{print}{diagnostics.compare_trafo}(x, ...)
}
\arguments{
\item{x}{an object of type \code{diagnostics.trafo_mod}}

\item{...}{additional arguments that are not used in this method}
}
\description{
prints objects to be shown in the summary function for objects of 
type \code{trafo_mod}
}
