% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.transformation.R
\name{print.transformation}
\alias{print.transformation}
\title{Print object of type transformation}
\usage{
\method{print}{transformation}(x, ...)
}
\arguments{
\item{x}{an object of type transformation
#@keywords internal}
}
\description{
Print object of type transformation
}
