% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.transformation.R
\name{print.transformation}
\alias{print.transformation}
\title{Print object of type transformation}
\usage{
\method{print}{transformation}(trans, ...)
}
\arguments{
\item{trans}{an object of type transformation with the estimates
#@keywords internal}
}
\description{
Print object of type transformation
}
