% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{reciprocal}
\alias{reciprocal}
\title{Reciprocal transformation for linear models}
\usage{
reciprocal(object)
}
\arguments{
\item{object}{an object of type lm.}
}
\value{
an object of class \code{trafo}.
}
\description{
The function transforms the dependent variable of a linear model using the 
Reciprocal transformation. The transformation parameter can either be 
estimated using different estimation methods or given.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Transform dependent variable 
reciprocal(object = lm_cars)
}
