% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/names_methods.R
\name{set_colnames}
\alias{set_colnames}
\alias{set_colnames.data.frame}
\alias{set_colnames.matrix}
\title{Set column names of a data frame or matrix.}
\usage{
set_colnames(x, ...)

\method{set_colnames}{data.frame}(x, col_names, ...)

\method{set_colnames}{matrix}(x, col_names, ...)
}
\arguments{
\item{x}{a data frame, tibble or matrix.}

\item{...}{extra parameters, currently none.}

\item{col_names}{a vector of unique column names.}
}
\value{
a data frame, tibble or matrix.
}
\description{
Sets column names of a data frame, tibble or matrix.
}
\details{
`set_colnames()` is a S3 generic function.
}
