% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trafo_mod.R
\name{summary.trafo_mod}
\alias{summary.trafo_mod}
\title{Summary for models with untransformed and transformed dependent variable}
\usage{
\method{summary}{trafo_mod}(object, ...)
}
\arguments{
\item{...}{additional arguments that are not used in this method}

\item{x}{an object of type \code{trafo_mod}}
}
\value{
an object of class \code{summary.transformation}
}
\description{
Information about the transformed data and model and components of an 
transformation object are extracted. The returned object is suitable for 
printing with the print.summary.transformation method.
}
\keyword{internal}
