% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trafo_mod.R
\name{summary.trafo_mod}
\alias{summary.trafo_mod}
\title{Summary for linear models with untransformed and transformed dependent 
variable}
\usage{
\method{summary}{trafo_mod}(object, ...)
}
\arguments{
\item{object}{an object of type \code{trafo_mod}}

\item{...}{additional arguments that are not used in this method}
}
\value{
an object of class \code{summary.trafo_mod}
}
\description{
The summary contains the summary for the untransformed and transformed model. 
The summary is based on the summary for objects of type \code{lm}. The 
returned object is suitable for printing with the print.summary.trafo_mod 
method.
}
