% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{summary.transformation}
\alias{summary.transformation}
\title{Summary transformations}
\usage{
\method{summary}{transformation}(x, ...)
}
\arguments{
\item{x}{an object of type \code{transformation}}
}
\value{
out an object of class \code{summary.transformation}
#@keywords internal
}
\description{
Computes objects to be shown in the summary function for objects of type \code{transformation}
}
