% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo.R
\docType{package}
\name{trafo}
\alias{trafo}
\alias{trafo-package}
\title{An R package supporting the selection of a suitable transformation}
\description{
Estimation, selection and comparison of several families of
transformations. The families of transformations included in the package are 
the Box-Cox transformation, the Manly transformation, the Yeo-Johnson 
transformation, the Modulus transformation, the Bickel-Docksum transformation 
and the dual transformation. The package simplifies to compare linear and 
linear mixed models with untransformed and transformed dependent variable. 
Furthermore, the package employs maximum likelihood approaches, skewness and 
divergence minimization to estimate the optimal transformation parameter.
}
\details{
An overview of all currently provided functions can be requested by
\code{library(help=trafo)}.
}
