% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo_lm.R
\name{trafo_lm}
\alias{trafo_lm}
\title{Fits transformed linear models}
\usage{
trafo_lm(object, trafo = "boxcoxshift", lambda = "estim", method = "ml",
  lambdarange = c(-2, 2), std = TRUE, custom_trafo = NULL)
}
\arguments{
\item{object}{an object of type \code{lm}.}

\item{trafo}{a character string. Different transformations can be used 
for transforming the dependent variable in a linear model: 
(i)  "bickeldoksum", (ii) "boxcoxshift", (iii) "boxcox", (iv) "dual", 
(v) "glog", (vi) "gpower", (vii) "log", (viii) "logshiftopt", (ix) "manly", 
(x) "modulus", (xi) "neglog", (xii) "reciprocal", (xiii) "yeojohnson".
Defaults to "boxcoxshift".}

\item{lambda}{either a character named "estim" if the optimal transformation
parameter should be estimated or a numeric value determining a given value 
for the transformation parameter. Defaults to "estim".}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Maximum likelihood approach ("ml"), (ii) Skewness minimization ("skew"),  
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cvm") or by Kullback-Leibler ("div.kl").
Defaults to "ml".}

\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter. 
Defaults to c(-2, 2).}

\item{std}{logical. If TRUE, the transformed model is returned based on the 
standardized transformation. Defaults to \code{TRUE}.}

\item{custom_trafo}{a list. The list has two elements where the first element 
is a function specifying the desired transformation and the second element is 
a function specifying the corresponding standardized transformation. 
Defaults to \code{NULL}.}
}
\value{
An object of class \code{trafo_mod}. Methods such as 
\code{\link{diagnostics.trafo_mod}}, \code{\link{print.trafo_mod}},
\code{\link{plot.trafo_mod}} and \code{\link{summary.trafo_mod}} can 
be used for this class.
}
\description{
Function \code{trafo_lm} fits linear models with transformed dependent 
variable. The main return are two \code{lm} objects where one is the 
untransformed linear model and the other one the transformed linear model.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Compare transformed models
trafo_lm(object = lm_cars, trafo = "bickeldoksum", method = "skew", 
lambdarange = c(1e-11, 2))
}
