% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo_lm.R
\name{trafo_lm}
\alias{trafo_lm}
\title{Fits transformed linear models}
\usage{
trafo_lm(object, trafo = "boxcox", lambda = "estim", method = "ml",
  lambdarange = c(-2, 2), std = FALSE, custom_trafo = NULL)
}
\arguments{
\item{object}{an object of type \code{lm}.}

\item{trafo}{a character string. Different transformations can be used 
for transforming the dependent variable in a linear model: 
(i)  "bickeldoksum", (ii) "boxcoxshift", (iii) "dual", (iv) "glog", 
(v) "gpower", (vi) "log", (vii) "logshiftopt", (viii) "manly", (ix) "modulus",
(x) "neglog", (xi) "reciprocal", (xii) "yeojohnson".
Defaults to "boxcoxshift".}

\item{lambda}{either a character named "estim" if the optimal transformation
parameter should be estimated or a numeric value determining a given 
transformation parameter. Defaults to "estim".}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Maximum likelihood approach ("ml"), (ii) Skewness minimization ("skew"),  
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cvm") or by Kullback-Leibler ("div.kl").
Defaults to "ml".}

\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter. 
Defaults to c(-2, 2).}

\item{std}{logical. If TRUE, the transformed model is returned based on the 
standardized transformation.}

\item{custom_trafo}{a list. If the customized transformation does 
not contain a transformation parameter the list has one element that is a 
function specifying the desired transformation. If the customized 
transformation contains a transformation parameter the list has two elements
where the first element is a function specifying the desired transformation
and the second element is a function specifying the corresponding standardized
transformation. Defaults to \code{NULL}.}
}
\value{
an object of class \code{trafo_mod}.
}
\description{
Function \code{trafo_lm} fits linear models with transformed dependent 
variable. The main return are two lm objects where one is the untransformed
linear model and the other one the transformed linear model.
}
\examples{
# Load data
data("cars", package = "datasets")

# Fit linear model
lm_cars <- lm(dist ~ speed, data = cars)

# Compare transformed models
trafo_lm(object = lm_cars, trafo = "bickeldoksum", method = "skew", 
lambdarange = c(1e-11, 2))
}
