% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo_lme.R
\name{trafo_lme}
\alias{trafo_lme}
\title{Fits transformed linear mixed models}
\usage{
trafo_lme(object, trafo, lambda = "estim", method, lambdarange,
  plotit = TRUE, std = FALSE)
}
\arguments{
\item{object}{an object of type lm or lme with the model to transform}

\item{trafo}{a character string. Different transformations can be used.}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter. 
(i) Maximum likelihood approaches: for linear models maximum likelihood ("ML")
and for linear mixed models restricted maximum likelihood ("reml"); 
(ii) Skewness minimizations: for linear models only skewness minimization 
("skew") and for linear mixed models also pooled skewness minimization; 
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cm") or by Kullback-Leibler ("div.kl") for both 
model types.}

\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter. 
Defaults to \code{c(-2, 2)} for the Box-Cox transformation.}
}
\value{
an object of class \code{transformation}
}
\description{
Function \code{trans_lm} fits linear mixed models with transformed dependent 
variables. The return are two lme objects where the first is the transformed
and the second the untransformed linear model.
}
\keyword{internal}
