% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo_lme.R
\name{trafo_lme}
\alias{trafo_lme}
\title{Fits transformed linear mixed models}
\usage{
trafo_lme(object, trafo, lambda = "estim", method, lambdarange,
  plotit = TRUE, std = FALSE)
}
\arguments{
\item{object}{an object of type lm or lme with the model to transform}

\item{trafo}{a character string. Different transformations can be used.}

\item{lambda}{either a character named "estim" if the optimal transformation
parameter should be estimated or a numeric value determining a given 
transformation parameter.}

\item{method}{a character string. Different estimation methods can be used 
for the estimation of the optimal transformation parameter: 
(i) Restricted maximum likelihood approach ("reml"), 
(ii) Skewness minimization ("skew") and pooled skewness minimization ("pskew"), 
(iii) Divergence minimization by Kolmogorov-Smirnoff ("div.ks"), 
by Cramer-von-Mises ("div.cm") or by Kullback-Leibler ("div.kl").}

\item{lambdarange}{a numeric vector with two elements defining an interval 
that is used for the estimation of the optimal transformation parameter.}

\item{plotit}{logical. If TRUE, a plot that illustrates the optimal 
transformation parameter or the given transformation parameter is returned.}

\item{std}{logical. If TRUE, the transformed model is returned based on the 
standardized transformation.}
}
\value{
an object of class \code{trafo_mod}.
}
\description{
Function \code{trans_lme} fits linear mixed models with one random effect 
and transformed dependent variable. The return are two lme objects where the 
first is the transformed and the second the untransformed linear mixed model.
}
\examples{
# Load data
data("eusilcA_Vienna")

# Fit linear mixed model
require(nlme)
lme_Vienna <- lme(eqIncome ~ eqsize + gender + cash + unempl_ben + age_ben +
rent + cap_inv + tax_adj + dis_ben + sick_ben + surv_ben + fam_allow + 
house_allow, random = ~ 1 | county, data = eusilcA_Vienna, 
na.action = na.omit)

# Get linear model with untransformed and transformed model
trafo_lme(object = lme_Vienna, trafo = "box.cox", method = "reml", 
lambdarange = c(0,2), plotit = TRUE, std = TRUE)
}
