% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woparam.lm.R
\name{woparam.lm}
\alias{woparam.lm}
\title{Transformations without parameter for linear models}
\usage{
\method{woparam}{lm}(object, trafo, custom_trafo = NULL)
}
\arguments{
\item{object}{an object of type lm.}

\item{trafo}{character that determines the transformation.}

\item{custom_trafo}{a function that specifies a transformation without 
paramater that needs to be estimated or given.}
}
\value{
an object of class \code{trafo}.
}
\description{
The function transforms the dependent variable of a linear model using 
different transformations without parameter.
}
\examples{
# Load data
data("eusilcA_Vienna")

# Fit linear model
lm_Vienna <- lm(eqIncome ~ eqsize + gender + cash + unempl_ben + age_ben + 
rent + cap_inv + tax_adj + dis_ben + sick_ben + surv_ben + 
fam_allow + house_allow, data = eusilcA_Vienna)

# Transform dependent variable using a maximum likelihood approach
bickeldoksum(object = lm_Vienna, lambda = "estim", method = "ml",
plotit = FALSE)
}
\references{
Battese, G.E., Harter, R.M. and Fuller, W.A. (1988). An Error-Components
Model for Predictions of County Crop Areas Using Survey and Satellite Data.
Journal of the American Statistical Association, Vol.83, No. 401, 28-36. \cr \cr
Gonzalez-Manteiga, W. et al. (2008). Bootstrap mean squared error of
a small-area EBLUP. Journal of Statistical Computation and Simulation,
78:5, 443-462.
}
