% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeojohnson.lm.R
\name{yeojohnson.lm}
\alias{yeojohnson.lm}
\title{Yeo-Johnson lm}
\usage{
\method{yeojohnson}{lm}(object, method, lambdarange = c(-2, 2), tol = 1e-04,
  ...)
}
\arguments{
\item{\dots}{additional arguments to be passed to the estimation function; see yeojohnsonEst()}

\item{lm}{object of type lm with the model to transform}
}
\value{
an object of class \code{transformation}; see yeojohnsonEst()
}
\description{
Yeo-Johnson estimation
}
\keyword{internal}
