% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeojohnson.R
\name{yeojohnson}
\alias{yeojohnson}
\title{Yeo-Johnson}
\usage{
yeojohnson(object, ...)
}
\arguments{
\item{object}{formula or fitted model; see below.}

\item{data}{an optional data frame, list or environment as in lm.}


\item{lambdarange}{range for the estimation parameter lambda - default c(-2, 2)}
}
\value{
An object of class \code{transformation} with the following arguments.

llike The value of profile log-likelihood at its maximum.

lambdahat The value of lambda for which the log-likelihood is maximized

logvector The profile log-likelihood evaluated at \code{lambdavector}.

lambdavector A sequence with optional values for \code{lambda}.

family Employed family of transformations.

yt Vector of the transformed response variable \code{y}.

zt Vector of the (normalized) transformed response variable \code{y}

modelt An object of type \code{lm} employing the transformed vector \code{zt} as the response variable
}

\references{
Yeo, I. and Johnson, R. 2000. A new family of power transformations to improve normality or symmetry. \emph{Biometrika}, \bold{87(4)}, 954--959.
}
\description{
Estimates the transformation parameter for the Yeo-Johnson transformation employing the profile log-likehood.
}
